[libunwind][PPC64] Added vector registers.
Summary:
The Registers_ppc64 class needed a couple of changes, both to accommodate the
new registers as well as to handle the overlaps of VS register set
without wasting space.

The save/restore code of V and VS registers was added.
As VS registers depend on the VMX extension, they are processed only if
VMX support is detected (_ARCH_PWR8 for now).

Note: on PowerPC64le (but not on x86), libcxx builts with no errors only
with clang, not with gcc. As libunwind tests depend on libcxx,
for now 'check-unwind' only works when LLVM/clang/libcxx/libunwind are
built with clang.
Building libunwind with gcc, in-tree without libcxx, works though, we
just can't run the tests in this case.

Differential Revision: https://reviews.llvm.org/D41906